+++ config/loom/index.ts
@@ -3,11 +3,13 @@ import {
   createComposedProjectPlugin,
   createProjectTestPlugin,
 } from '@shopify/loom';
-import {babel} from '@shopify/loom-plugin-babel';
-import {packageBuild} from '@shopify/loom-plugin-package-build';
+import {buildLibrary, babel} from '@shopify/loom-plugin-build-library';
+
+// Needed so TS realises what configuration hooks are provided by Jest
 import type {} from '@shopify/loom-plugin-jest';
 
 export function quiltPackage({
+  isIsomorphic = true,
   jestEnv = 'jsdom',
   jestTestRunner = 'jest-circus',
   polyfill = true,
@@ -16,7 +18,19 @@ export function quiltPackage({
   // occur we need to turn some options off
   const polyfillOptions = polyfill ? {} : {useBuiltIns: false, corejs: false};
 
+  const targets = isIsomorphic
+    ? 'extends @shopify/browserslist-config, node 12.14.0'
+    : 'node 12.14.0';
+
   return createComposedProjectPlugin<Package>('Quilt.Package', [
+    buildLibrary({
+      targets,
+      commonjs: true,
+      esmodules: true,
+      esnext: true,
+      rootEntrypoints: true,
+      jestTestEnvironment: jestEnv,
+    }),
     babel({
       config: {
         presets: [
@@ -27,21 +41,16 @@ export function quiltPackage({
         ],
       },
     }),
-    packageBuild({
-      nodeTargets: 'node 12.14.0',
-      browserTargets: 'extends @shopify/browserslist-config',
-    }),
     createProjectTestPlugin('Quilt.PackageTest', ({hooks}) => {
       hooks.configure.hook((hooks) => {
-        hooks.jestEnvironment?.hook(() => jestEnv);
         hooks.jestTestRunner?.hook(() => jestTestRunner);
 
-        hooks.jestTransforms?.hook((transforms) => ({
+        hooks.jestTransform?.hook((transforms) => ({
           ...transforms,
           '\\.(gql|graphql)$': 'jest-transform-graphql',
         }));
 
-        hooks.jestWatchIgnore?.hook((patterns) => [
+        hooks.jestWatchPathIgnorePatterns?.hook((patterns) => [
           ...patterns,
           '<rootDir>/.*/tests?/.*fixtures',
         ]);