+++ packages/react-async/loom.config.ts
@@ -1,10 +1,9 @@
-import {createPackage, Runtime} from '@shopify/loom';
+import {createPackage} from '@shopify/loom';
 
 import {quiltPackage} from '../../config/loom';
 
 export default createPackage((pkg) => {
-  pkg.runtimes(Runtime.Browser, Runtime.Node);
-  pkg.entry({root: './src/index'});
-  pkg.entry({name: 'testing', root: './src/testing'});
+  pkg.entry({root: './src/index.ts'});
+  pkg.entry({name: 'testing', root: './src/testing.tsx'});
   pkg.use(quiltPackage());
 });