+++ packages/storybook-a11y-test/CHANGELOG.md
@@ -11,6 +11,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Updated puppeteer to 13.1.3 ([#2149](https://github.com/Shopify/quilt/pull/2149))
 - Reorder exports map to prioritize the `esnext` condition. [[#2148](https://github.com/Shopify/quilt/pull/2148)]
+- Update Loom build config. [[#2150](https://github.com/Shopify/quilt/pull/2150)]
 
 ## 0.4.3 - 2022-01-10